pyproject.toml
[project]
name = "ieee-chess"
version = "0.1.0"
description = "puzzle theme embedding"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chess>=1.11.2",
    "numpy>=2.2.6",
    "onnx>=1.18.0",
    "onnx2torch>=1.5.15",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
    "zstandard>=0.23.0",
]
